requirements.txt
flask
flask-cors
pinecone-client
python-dotenv
langchain
openai
PyPDF2
